Fixed docker not building.
The problem was caused because one of the dependencies didn't have a fixed version. so there was a conflict on php version required for the project and the dependency, I kinda solve it on a hacky way, but didn't want to use a repository not maintained by the official team.